Estilos basicos de la dashboard
