+++ src/components/CustomPackage/CustomPackage.css
@@ -40,7 +40,6 @@
     border-radius: 5px;
     color: #fff;
     cursor: pointer;
-     
     
     }
